+++ core/components/seosuite/lexicon/en/default.inc.php
@@ -109,3 +109,13 @@ $_lang['seosuite.time_months']                                  = '[[+months]] m
 $_lang['seosuite.time_to_long']                                 = 'More then a half year ago';
 
 $_lang['seosuite.widget_desc']                                  = 'Here you can view the 10 most recently added 404 urls. To view and manage all your 404 urls, please visit the <a href="[[++manager_url]]?a=home&amp;namespace=seosuite">SEO Suite manager page.</a>';
+
+$_lang['seosuite.migration']                                    = 'Migration';
+$_lang['seosuite.migration_desc']                               = 'Upgrading from SEO Suite V1, SEO Pro or SEO Tab requires a data migration. Here you can run the migrations for these extras.';
+$_lang['seosuite.migration.seosuitev1.results']                 = '<b>[[+redirects]]</b> Redirects and <b>[[+urls]]</b> urls for SEO Suite V1 found.';
+$_lang['seosuite.migration.seosuitev1.empty']                   = '<i>No data found for SEO Suite V1 migration.</i>';
+$_lang['seosuite.migration.seopro.results']                     = '<b>[[+count]]</b> SEO Pro keyword resources found.';
+$_lang['seosuite.migration.seopro.empty']                       = '<i>No data found for SEO Pro migration.</i>';
+$_lang['seosuite.migration.seotab.results']                     = '<b>[[+urls]]</b> Redirects and <b>[[+resources]]</b> resources for SEO Tab found.';
+$_lang['seosuite.migration.seotab.empty']                       = '<i>No data found for SEO Tab migration.</i>';
+$_lang['seosuite.migration.migrate']                            = 'Migrate';